requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
langchain-chroma
chromadb
sentence-transformers
pandas
openpyxl
streamlit
python-dotenv